+++ lib/karafka/connection/client.rb
@@ -23,7 +23,7 @@ module Karafka
       # @yieldparam [Array<Kafka::FetchedMessage>] kafka fetched messages
       # @note This will yield with raw messages - no preprocessing or reformatting.
       def fetch_loop
-        settings = ConfigAdapter.consuming(consumer_group)
+        settings = ApiAdapter.consuming(consumer_group)
 
         if consumer_group.batch_fetching
           kafka_consumer.each_batch(*settings) { |batch| yield(batch.messages) }
@@ -69,7 +69,7 @@ module Karafka
         kafka_consumer.pause(
           Karafka::App.config.topic_mapper.outgoing(topic),
           partition,
-          ConfigAdapter.pausing(consumer_group)
+          ApiAdapter.pausing(consumer_group)
         )
       end
 
@@ -92,10 +92,10 @@ module Karafka
       #   that is set up to consume from topics of a given consumer group
       def kafka_consumer
         @kafka_consumer ||= kafka.consumer(
-          *ConfigAdapter.consumer(consumer_group)
+          *ApiAdapter.consumer(consumer_group)
         ).tap do |consumer|
           consumer_group.topics.each do |topic|
-            consumer.subscribe(*ConfigAdapter.subscription(topic))
+            consumer.subscribe(*ApiAdapter.subscription(topic))
           end
         end
       rescue Kafka::ConnectionError
@@ -111,7 +111,7 @@ module Karafka
       # @note We don't cache it internally because we cache kafka_consumer that uses kafka
       #   object instance
       def kafka
-        Kafka.new(*ConfigAdapter.client(consumer_group))
+        Kafka.new(*ApiAdapter.client(consumer_group))
       end
     end
   end